refactor: common 패키지에 공통 헤더 분리
