+++ app/src/main/java/nextstep/jwp/controller/AbstractController.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.model.httpMessage.HttpMethod;
+import nextstep.jwp.model.httpMessage.request.HttpMethod;
 import nextstep.jwp.model.httpMessage.request.HttpRequest;
 import nextstep.jwp.model.httpMessage.response.HttpResponse;
