+++ app/src/main/java/nextstep/jwp/model/httpMessage/common/HttpHeaderType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model.httpMessage;
+package nextstep.jwp.model.httpMessage.common;
 
 import java.util.Arrays;
 
@@ -17,14 +17,14 @@ public enum HttpHeaderType {
                 .anyMatch(type -> type.value.equals(name));
     }
 
-    public String value() {
-        return value;
-    }
-
     public static HttpHeaderType of(String value) {
         return Arrays.stream(values())
                 .filter(headerType -> headerType.value.equals(value))
                 .findAny()
                 .orElseThrow(() -> new IllegalArgumentException("해당 헤더 타입이 존재하지 않습니다."));
+    }
+
+    public String value() {
+        return value;
     }
 }
