+++ app/src/main/java/nextstep/jwp/model/httpMessage/request/HttpRequest.java
@@ -1,6 +1,5 @@
 package nextstep.jwp.model.httpMessage.request;
 
-import nextstep.jwp.model.httpMessage.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.thymeleaf.util.StringUtils;
@@ -13,8 +12,8 @@ import java.nio.charset.StandardCharsets;
 import java.util.Map;
 import java.util.Objects;
 
-import static nextstep.jwp.model.httpMessage.ContentType.FORM;
-import static nextstep.jwp.model.httpMessage.HttpHeaderType.CONTENT_LENGTH;
+import static nextstep.jwp.model.httpMessage.common.ContentType.FORM;
+import static nextstep.jwp.model.httpMessage.common.HttpHeaderType.CONTENT_LENGTH;
 
 public class HttpRequest {
     private static final Logger log = LoggerFactory.getLogger(HttpRequest.class);
