+++ app/src/main/java/nextstep/jwp/model/httpMessage/request/RequestHeaderType.java
@@ -15,10 +15,6 @@ public enum RequestHeaderType {
         this.value = value;
     }
 
-    public String value() {
-        return value;
-    }
-
     public static boolean contains(String name) {
         return Arrays.stream(values())
                 .anyMatch(type -> type.value.equals(name));
@@ -29,5 +25,9 @@ public enum RequestHeaderType {
                 .filter(type -> type.value.equals(value))
                 .findAny()
                 .orElseThrow(() -> new IllegalArgumentException("Not found request Header type (input : " + value + ")"));
+    }
+
+    public String value() {
+        return value;
     }
 }
