+++ app/src/test/java/nextstep/jwp/model/RequestLineTest.java
@@ -3,8 +3,8 @@ package nextstep.jwp.model;
 import nextstep.jwp.model.httpMessage.request.RequestLine;
 import org.junit.jupiter.api.Test;
 
-import static nextstep.jwp.model.httpMessage.HttpMethod.GET;
-import static nextstep.jwp.model.httpMessage.HttpMethod.POST;
+import static nextstep.jwp.model.httpMessage.request.HttpMethod.GET;
+import static nextstep.jwp.model.httpMessage.request.HttpMethod.POST;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class RequestLineTest {
